security improved on accessing not allowed pages, comment mail bug fix, getting media stream from camera

+++ server/deletefromgallery.php
@@ -3,17 +3,22 @@ require_once('../config/connect.php');
 session_start();
 if (isset($_POST['delete']) && $_POST['delete'] != '' && isset($_POST['name']) && $_POST['name'] != '' && isset($_SESSION['login']) && $_SESSION['login'] != '') {
     try {
-        $stmt = $db->prepare('DELETE FROM gallery WHERE id = :id');
+        $stmt = $db->prepare('DELETE FROM gallery WHERE id = :id AND login = :log');
         $stmt->bindParam(':id', $_POST['delete'], PDO::PARAM_INT);
+        $stmt->bindParam(':log', $_SESSION['login']);
         $stmt->execute();
-        $stmt = $db->prepare('DELETE FROM comments WHERE gallery_id = :id');
-        $stmt->bindParam(':id', $_POST['delete'], PDO::PARAM_INT);
-        $stmt->execute();
-        $stmt = $db->prepare('DELETE FROM likes WHERE id_gallery = :id');
-        $stmt->bindParam(':id', $_POST['delete'], PDO::PARAM_INT);
-        $stmt->execute();
-        unlink('../img/' . $_POST['name']);
-        echo "success : picture deleted";
+        if ($stmt->rowCount() > 0) {
+            $stmt = $db->prepare('DELETE FROM comments WHERE gallery_id = :id');
+            $stmt->bindParam(':id', $_POST['delete'], PDO::PARAM_INT);
+            $stmt->execute();
+            $stmt = $db->prepare('DELETE FROM likes WHERE id_gallery = :id');
+            $stmt->bindParam(':id', $_POST['delete'], PDO::PARAM_INT);
+            $stmt->execute();
+            unlink('../img/' . $_POST['name']);
+            echo "success : picture deleted";
+        } else {
+            echo "error : picture not deleted";
+        }
     } catch( PDOException $Exception ) {
         echo 'Error: '.$Exception->getMessage();
         die();
